Replace MiniListItem_t with ListItem_t
MiniListItem_t is intended to be the last element of a linked list. It
saves space by having a payload but no `next' pointer, since there is no
next element. CBMC gets confused when it sees a list of ListItem_ts with
a smaller MiniListItem_t at the end, so this commit makes the last
element of lists a ListItem_t instead.

This patch should be removed soon: It is for a CBMC issue being fixed now.